Move scroll_label.py into RimSort.models
